+++ cc-service-api/src/main/java/com/jtmc/apps/reforma/api/v1/user/UserResponse.java
@@ -6,6 +6,15 @@ public class UserResponse {
     private List<Integer> roles;
     private int defaultBranch;
     private String username;
+    private int userId;
+
+    public int getUserId() {
+        return userId;
+    }
+
+    public void setUserId(int userId) {
+        this.userId = userId;
+    }
 
     public List<Integer> getRoles() {
         return roles;