+++ cc-service-impl/src/main/java/com/jtmc/apps/reforma/repository/PeriodConfirmRepository.java
@@ -2,6 +2,7 @@ package com.jtmc.apps.reforma.repository;
 
 import com.google.inject.Inject;
 import com.jtmc.apps.reforma.domain.PeriodDetails;
+import com.jtmc.apps.reforma.repository.exception.RepositoryException;
 import com.jtmc.apps.reforma.repository.mapper.PeriodDetailsMapper;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
@@ -23,17 +24,17 @@ public class PeriodConfirmRepository {
             return mapper.insert(periodDetails);
         } catch (Exception ex) {
            logger.error("{}", ex);
-           throw  ex;
+           throw new RepositoryException("Error in PeriodConfirmRepository", 500);
         }
     }
 
-    public Optional<PeriodDetails> selectOne(int branchId, int periodId) {
+    public Optional<PeriodDetails> selectOne(int branchId, int periodId, int confirmedBy) {
          try(SqlSession session = sqlSessionFactory.openSession()) {
              PeriodDetailsMapper mapper = session.getMapper(PeriodDetailsMapper.class);
-            return mapper.selectByPrimaryKey(periodId, branchId);
+            return mapper.selectByPrimaryKey(branchId, periodId, confirmedBy);
         } catch (Exception ex) {
            logger.error("{}", ex);
-           throw  ex;
+           throw new RepositoryException("Error in PeriodConfirmRepository", 500);
         }
     }
 }